requirements.txt
numpy==1.21.4
matplotlib.pyplot==3.4.3
tensorflow==2.7.0
opencv-python==4.5.4.58